requirements.txt

python-dotenv==1.1.1
pytest==8.4.1
requests==2.32.4
allure-pytest==2.14.3
jsonschema==4.24.0
allure-python-commons~=2.14.3